Show edit pas part of the modal
