+++ src/app/tidbits/edit/[[...byteId]]/page.tsx
@@ -1,181 +1,33 @@
 'use client';
 
 import withSpace from '@/app/withSpace';
-import Block from '@/components/app/Block';
-import DeleteConfirmationModal from '@/components/app/Modal/DeleteConfirmationModal';
-import AddByteQuestionsUsingAIButton from '@/components/bytes/Create/AddByteQuestionsUsingAIButton';
-import { CreateByteUsingAIModal } from '@/components/bytes/Create/CreateByteUsingAIModal';
-import { EditByteType } from '@/components/bytes/Edit/editByteHelper';
-import EditByteStepper from '@/components/bytes/Edit/EditByteStepper';
-import { useEditByte } from '@/components/bytes/Edit/useEditByte';
-import Button from '@/components/core/buttons/Button';
-import { EllipsisDropdownItem } from '@/components/core/dropdowns/EllipsisDropdown';
-import PrivateEllipsisDropdown from '@/components/core/dropdowns/PrivateEllipsisDropdown';
-import Input from '@/components/core/input/Input';
-import PageLoading from '@/components/core/loaders/PageLoading';
-import PageWrapper from '@/components/core/page/PageWrapper';
-import TextareaArray from '@/components/core/textarea/TextareaArray';
-import { SpaceWithIntegrationsFragment, useDeleteByteMutation } from '@/graphql/generated/generated-types';
-import SingleCardLayout from '@/layouts/SingleCardLayout';
-import { TidbitShareSteps } from '@/types/deprecated/models/enums';
-import { ByteErrors } from '@/types/errors/byteErrors';
-import { useSession } from 'next-auth/react';
-import { router } from 'next/client';
-import Link from 'next/link';
-import { useEffect, useState } from 'react';
+import EditByteView from '@/components/bytes/Edit/EditByteView';
+import FullScreenModal from '@/components/core/modals/FullScreenModal';
+import { SpaceWithIntegrationsFragment } from '@/graphql/generated/generated-types';
+import { useRouter } from 'next/navigation';
+import React from 'react';
 
 function EditByte(props: { space: SpaceWithIntegrationsFragment; params: { byteId?: string[] } }) {
   const { space, params } = props;
   const byteId = params.byteId ? params.byteId[0] : null;
 
-  const { byteUpserting, byteLoaded, byteRef: byte, byteErrors, handleByteUpsert, initialize, updateByteFunctions } = useEditByte(space, byteId || null);
-  const { data: session } = useSession();
-  const inputError = (field: keyof ByteErrors): string => {
-    const error = byteErrors?.[field];
-    return error ? error.toString() : '';
-  };
-
-  useEffect(() => {
-    initialize();
-  }, [byteId]);
-
-  const [showAIGenerateModel, setShowAIGenerateModel] = useState(false);
-  const threeDotItems: EllipsisDropdownItem[] = [{ label: 'Delete', key: 'delete' }];
-
-  const [showDeleteModal, setShowDeleteModal] = useState(false);
-
-  const [deleteByteMutation] = useDeleteByteMutation();
+  const router = useRouter();
+  function onClose() {
+    router.push(`/tidbits`);
+  }
 
   return (
-    <PageWrapper>
-      <SingleCardLayout>
-        <div className="px-4 mb-4 md:px-0 flex justify-between">
-          <Link href={byteId ? `/tidbits/view/${byteId}/0` : `/tidbits`} className="text-color">
-            <span className="mr-1 font-bold">&#8592;</span>
-            {byteId ? byte.name : 'Back to Bytes'}
-          </Link>
-          <div>
-            {!byteId && <Button onClick={() => setShowAIGenerateModel(true)}>Create with AI</Button>}
-            <AddByteQuestionsUsingAIButton
-              byte={byte}
-              onNewStepsWithQuestions={(newSteps) => {
-                updateByteFunctions.includeSteps(newSteps);
-              }}
-            />
-            {byteId && (
-              <PrivateEllipsisDropdown
-                items={threeDotItems}
-                onSelect={(key) => {
-                  if (key === 'delete') {
-                    setShowDeleteModal(true);
-                  }
-                }}
-                className="ml-4 z-50"
-              />
-            )}
-          </div>
-        </div>
-
-        {byteLoaded ? (
-          <div className="pb-10">
-            <Block title="Basic Info" className="mt-4">
-              <div className="mb-2">
-                <Input modelValue={byte.name} error={inputError('name')} maxLength={32} onUpdate={(e) => updateByteFunctions.updateByteField('name', e)}>
-                  Name *
-                </Input>
-                <Input
-                  modelValue={byte.content}
-                  error={inputError('content') ? 'Excerpt is required and should be less than 64 characters long' : ''}
-                  placeholder="byte.create.excerpt"
-                  maxLength={64}
-                  onUpdate={(e) => updateByteFunctions.updateByteField('content', e)}
-                >
-                  Excerpt *
-                </Input>
-
-                <TextareaArray
-                  label="Admins"
-                  id="admins"
-                  modelValue={byte.admins}
-                  placeholder={`0x8C28Cf33d9Fd3D0293f963b1cd27e3FF422B425c\n0xeF8305E140ac520225DAf050e2f71d5fBcC543e7`}
-                  className="input w-full text-left"
-                  onUpdate={(e) => updateByteFunctions.updateByteField('admins', e)}
-                />
-
-                <Input
-                  modelValue={byte.videoUrl}
-                  placeholder="byte.create.videoURL"
-                  maxLength={1024}
-                  onUpdate={(e) => updateByteFunctions.updateByteField('videoUrl', e)}
-                >
-                  Video URL
-                </Input>
-
-                <TextareaArray
-                  label="Tags"
-                  id="tags"
-                  modelValue={byte.tags}
-                  placeholder={`tag1\ntag2`}
-                  className="input w-full text-left"
-                  onUpdate={(e) => updateByteFunctions.updateByteField('tags', e)}
-                />
-
-                <div className="mt-4">
-                  <Input
-                    modelValue={byte.priority}
-                    maxLength={500}
-                    onUpdate={(e) => updateByteFunctions.updateByteField('priority', e)}
-                    className="edit-timeline-inputs"
-                    number
-                  >
-                    Priority
-                  </Input>
-                </div>
-              </div>
-            </Block>
-            <Block title="Byte Steps" slim={true}>
-              <div className="mt-4">
-                <EditByteStepper space={space} byte={byte} byteErrors={byteErrors} updateByteFunctions={updateByteFunctions} />
-              </div>
-            </Block>
-
-            <div className="flex">
-              <Button
-                onClick={handleByteUpsert}
-                loading={byteUpserting}
-                disabled={!byteLoaded || byteUpserting}
-                className="ml-2 block w-full"
-                variant="contained"
-                primary
-              >
-                Upsert
-              </Button>
-            </div>
-          </div>
-        ) : (
-          <PageLoading />
-        )}
-      </SingleCardLayout>
-      <CreateByteUsingAIModal
-        open={showAIGenerateModel}
-        onClose={() => setShowAIGenerateModel(false)}
-        onGenerateByte={(generated: EditByteType) => {
-          updateByteFunctions.setByte(generated);
-        }}
-      />
-      {showDeleteModal && (
-        <DeleteConfirmationModal
-          title={'Delete Byte'}
-          open={showDeleteModal}
-          onClose={() => setShowDeleteModal(true)}
-          onDelete={async () => {
-            await deleteByteMutation({ variables: { spaceId: space.id, byteId: byteId! } });
-            setShowDeleteModal(false);
-            router.push(`/tidbits`);
+    <FullScreenModal open={true} onClose={onClose} title={'Create Tidbit'}>
+      <div className="text-left">
+        <EditByteView
+          space={space}
+          byteId={byteId || undefined}
+          onUpsert={async (byteId) => {
+            router.push(`/tidbits/view/${byteId}`);
           }}
         />
-      )}
-    </PageWrapper>
+      </div>
+    </FullScreenModal>
   );
 }
